Refactor level sensor and SHT20 data handling: change level data type to uint16_t, simplify SHT20 read function, and update JSON formatting. Adjust system registers for threshold alarm functionality and modify Modbus server unit ID.

+++ P2_RS485_Controllable_Breaker/devices/sensors/level_sensor/include/level_sensor.h
@@ -8,11 +8,11 @@
 #define LEVEL_SENSOR_LEVEL_ADDR 0x0001
 
 struct level_sensor_param {
-	float level;
+	uint16_t level;
 };
 
 int level_sensor_read_data(uint8_t unit_id, struct level_sensor_param *data);
-int level_sensor_get_level(uint8_t unit_id, float *level);
+int level_sensor_get_level(uint8_t unit_id, uint16_t *level);
 int level_sensor_get_data(uint8_t unit_id, char *buffer, size_t buffer_size);
 bool level_sensor_validate_data(const struct level_sensor_param *data);
 int level_sensor_format_json(const struct level_sensor_param *data,
